Remove extra host indirection for webviews
No longer needed now that we only have iframe based webviews
