+++ src/vs/workbench/contrib/webview/browser/pre/index.html
@@ -13,7 +13,7 @@
 </head>
 
 <body style="margin: 0; overflow: hidden; width: 100%; height: 100%" role="document">
-	<script async type="module" src="host.js"></script>
+	<script async type="module" src="main.js"></script>
 </body>
 
 </html>
